# api/openapi.yaml
openapi: 3.0.3

info:
  title: Learn The Grammar API
  version: 0.1.0
  description: >
    Contract for the backend API consumed by the frontend. This is the
    single source of truth: server types (backend/internal/api) and the
    frontend client (frontend/src/api) are generated from this file.
    Infrastructure-only endpoints (e.g. /health) are intentionally not
    part of this contract.

paths:
  /api/hello:
    get:
      operationId: getHello
      summary: Returns a greeting message.
      responses:
        default:
          $ref: "#/components/responses/Error"
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HelloResponse"

  /api/me:
    get:
      operationId: getMe
      summary: Returns the current session's user ID.
      description: >
        Every request automatically carries or receives a session
        (see the Set-Cookie response header on first visit) identifying
        a user. There is no login yet, so this ID is not backed by any
        registered account — it simply identifies the current browser
        across requests.
      responses:
        default:
          $ref: "#/components/responses/Error"
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"

  /api/settings:
    get:
      operationId: getSettings
      summary: Returns the current learning settings.
      responses:
        default:
          $ref: "#/components/responses/Error"
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SettingsResponse"
    put:
      operationId: putSettings
      summary: Saves the learning settings and completes onboarding.
      description: >
        Onboarding, once completed, stays completed: a later save never
        un-completes it.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveSettingsRequest"
      responses:
        default:
          $ref: "#/components/responses/Error"
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SettingsResponse"

components:
  responses:
    Error:
      description: Request failed.
      headers:
        X-Request-ID:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      required: [code, message, request_id]
      properties:
        code:
          type: string
        message:
          type: string
        request_id:
          type: string
    HelloResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string

    MeResponse:
      type: object
      required:
        - user_id
      properties:
        user_id:
          type: string
          format: uuid

    Level:
      type: string
      enum: [A1, A2, B1]

    Topic:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string

    SettingsResponse:
      type: object
      required: [level, onboarding_completed, topics]
      properties:
        level:
          $ref: "#/components/schemas/Level"
        onboarding_completed:
          type: boolean
        topics:
          type: array
          items:
            $ref: "#/components/schemas/Topic"

    SaveSettingsRequest:
      type: object
      required: [level]
      properties:
        level:
          $ref: "#/components/schemas/Level"